% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{APF}
\alias{APF}
\title{Auxiliary Particle Filter}
\usage{
APF(obs, num_particles, param)
}
\arguments{
\item{obs}{A vector of observations to be filtered.}

\item{num_particles}{A positive integer number of particles to be used
in the simulation.}

\item{param}{A vector of model parameters.}
}
\value{
A list containing a sample from the empirical distribution; the
approximated marginal log-likelihood of the data; the filtered states.
}
\description{
This function is currently limited to the stochastic volatility model.
}
