% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{APF_Cpp}
\alias{APF_Cpp}
\title{Auxiliary Particle Filter}
\usage{
APF_Cpp(obs, num_particles, param)
}
\arguments{
\item{obs}{A vector of observations to be filtered.}

\item{num_particles}{A positive integer number of particles to be used
in the simulation.}

\item{param}{A vector of model parameters (alpha, beta, sigma).}
}
\value{
A list containing a sample from the empirical distribution; the
approximated marginal log-likelihood of the data; the sampled particles 
and their associated weights.
}
\description{
This function is currently limited to the stochastic volatility model.
}
