% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Bootstrap_SV_C}
\alias{Bootstrap_SV_C}
\title{Bootstrap class for Stochastic Volatility (SV) model}
\description{
Initialised with a data vector of observations (Yt), an estimate mean and 
standard deviation parameter and rho controls the autocorrelation.
}
\details{
The object defines methods to sample from the initial distribution M0 and 
the subsequent transition kernels Mt and calculate the log probability from 
the potential function Gt.
}
\section{Fields}{

\describe{
\item{\code{data}}{vector of observations}

\item{\code{mu}}{estimate of mean of latent var (float)}

\item{\code{sigma}}{estimate of sd of latent}

\item{\code{rho}}{autocorrelation parameter}
}}

