% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BSF}
\alias{BSF}
\title{Bootstrap Particle Filter}
\usage{
BSF(obs, N, param)
}
\arguments{
\item{obs}{A vector of observations to be filtered.}

\item{N}{The number of particles to be used in the filter.}

\item{param}{The model parameters to be used within the filter. These are
passed in the form c(alpha, beta, sigma).}
}
\value{
A list containing the filtered states; a sample from the posterior
distribution; the log-marginal distribution of the observations.
}
\description{
Bootstrap Particle Filter
}
