% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eff_particle_no}
\alias{eff_particle_no}
\title{Effective Particle Number}
\usage{
eff_particle_no(w)
}
\description{
Uses the weight vector input to compute an estimate of effective particle number for use in particle filters.
}
\section{Fields}{

\describe{
\item{\code{w}}{vector of weights}
}}

