% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pmmh1}
\alias{pmmh1}
\title{Pseudo-Marginal Metropolis Hastings}
\usage{
pmmh1(thetastart, niter, N, y, burnin, alphastart, beta)
}
\arguments{
\item{thetastart}{The starting values for theta (sigma?).}

\item{niter}{The length of the chains.}

\item{N}{The number of particles to used in the marginal likelihood
approximation.}

\item{y}{A vector of observations.}

\item{burnin}{The size of the burn-in period.}

\item{alphastart}{Initial value for alpha parameter.}

\item{beta}{The true value of beta from which the data is generated.}
}
\description{
Using a beta distribution as the proposal and gamma priors on alpha and
Sigma.
}
