
%%% $Id: print.Rd,v 1.2 2008/02/04 21:40:10 goswami Exp $

\name{print}

\alias{print.SMC}

\title{The printing family of functions}

\description{
  The printing family of functions for this package.
}

\usage{
\method{print}{SMC}(x, \dots)
}

\arguments{
  \item{x}{an object inheriting from class \code{SMC} (generated by
    functions \code{particleFilter}, \code{auxiliaryParticleFilter} and
    \code{sequentialMonteCarlo}).}

  \item{\dots}{optional arguments passed to \code{print.default}; see
    its documentation.}
  
}

\author{Gopi Goswami \email{goswami@stat.harvard.edu}}

\seealso{\code{\link{particleFilter}},
  \code{\link{auxiliaryParticleFilter}},
  \code{\link{sequentialMonteCarlo}}}

\keyword{print}
