% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_SV.R
\name{systematic_resampling}
\alias{systematic_resampling}
\title{Systematic resampling algorithm}
\usage{
systematic_resampling(W)
}
\description{
Samples a single uniform random variable U and then assigns values U(n) = (n - 1 + U)/N.
Used to generate a vector of integers to index the particles.
}
